Update vignette and examples
I believe this commit wraps up the last suggestions made by
@zachary-foster.

I kind of half-assed the 'contribution guidelines' requirement, with the
following very minimal entry in the README:

```
 # Contributing

 We welcome any contributions!

  * By participating in this project you agree to abide by the terms outlined in
    the [Contributor Code of Conduct](CONDUCT.md).
```

Also I have not fixed the issue with the two unnamed loading bars. This
is probably an issue with `taxizedb`, not `onekp`, so is not something I
can fix in this package.

My discussion of the intended user-base in the README is still a little
weak.

Other than these issues, I believe I fixed every issue raised without
any serious disagreement.